Refactor scraper architecture; implement TCGPlayerClient, card parsing, and sealed price fetching; enhance Excel export functionality
